channelUser in context
