+++ vivid-client/src/hooks/useMessages.ts
@@ -1,4 +1,5 @@
 import React from 'react';
+import { UserContext } from './useUser';
 import { UsersContext } from './useUsers';
 import { SocketContext } from './useWebsocket';
 
@@ -11,6 +12,12 @@ export function useMessages(channel: string) {
   const { messages, setChannelMessages, getChannelMessages } =
     React.useContext(MessageContext);
   const { addUser, getUser, users } = React.useContext(UsersContext);
+
+  const userData = React.useContext(UserContext);
+  const [currentChannelUser, setCurrentChannelUser] = React.useState<{
+    user: any;
+    tag: any;
+  }>({ user: null, tag: null });
 
   const [error, setError] = React.useState(false);
   const [isLoading, setLoading] = React.useState(true);
@@ -52,6 +59,32 @@ export function useMessages(channel: string) {
       });
   }
 
+  function getRole(userId: string) {
+    if (!userId) return null;
+    let tag = null;
+    const channelUser = channelInfo?.joined_users.find(
+      (v: any) => v.user?.id === userId || v.user === userId,
+    );
+    if (channelUser && channelUser.is_mod) tag = 'mod';
+    if (userId === channelInfo?.owner) tag = 'owner';
+    return tag;
+  }
+
+  React.useEffect(() => {
+    if (!channelInfo || !userData.user?.id) {
+      setCurrentChannelUser({ user: null, tag: null });
+      return;
+    }
+    const n = {
+      user: channelInfo?.joined_users?.find(
+        (v: any) =>
+          v.user?.id === userData.user?.id || v.user === userData.user?.id,
+      ),
+      tag: getRole(userData.user?.id),
+    };
+    setCurrentChannelUser(n);
+  }, [channelInfo, userData]);
+
   React.useEffect(() => {
     requestMessages();
   }, [channel]);
@@ -92,6 +125,8 @@ export function useMessages(channel: string) {
     channelInfo,
     getUser,
     users,
+    currentChannelUser,
+    getRole,
     messageState: {
       error,
       loading: isLoading,
