+++ vivid-client/src/views/ChannelView.tsx
@@ -11,37 +11,9 @@ export function ChannelView() {
   const scrollEl = React.useRef(null);
   const { id }: any = useParams();
   const messageData = useMessages(id);
+  const { getRole, currentChannelUser } = messageData;
   const userData = React.useContext(UserContext);
   const [reducedMessages, setReducedMessages] = React.useState<any[]>([]);
-  const [currentUser, setCurrentUser] = React.useState<{
-    user?: any | null;
-    tag?: string | null;
-  }>({ user: null, tag: null });
-
-  function getRole(userId: string) {
-    let tag = null;
-    const channelUser = messageData?.channelInfo?.joined_users.find(
-      (v: any) => v.user?.id === userId || v.user === userId,
-    );
-    if (channelUser && channelUser.is_mod) tag = 'mod';
-    if (userId === messageData?.channelInfo?.owner) tag = 'owner';
-    return tag;
-  }
-
-  React.useEffect(() => {
-    if (!messageData?.channelInfo || !userData.user?.id) {
-      setCurrentUser({ user: null, tag: null });
-      return;
-    }
-    const n = {
-      user: messageData?.channelInfo?.joined_users?.find(
-        (v: any) =>
-          v.user?.id === userData.user?.id || v.user === userData.user?.id,
-      ),
-      tag: currentUser ? getRole(userData.user?.id) : null,
-    };
-    setCurrentUser(n);
-  }, [messageData.channelInfo, userData]);
 
   React.useEffect(() => {
     const cur: any = scrollEl?.current;
@@ -110,7 +82,8 @@ export function ChannelView() {
     <MainLayout
       title={
         (messageData.messageState.done ? messageData.channelInfo.title : '‎') +
-        (currentUser.tag && ['owner', 'mod'].includes(currentUser.tag)
+        (currentChannelUser.tag &&
+        ['owner', 'mod'].includes(currentChannelUser.tag)
           ? 'Edit channel'
           : '')
       }
